backend/TravelSip/requirements.txt
Django==4.2.6
django-cors-headers
django-phonenumber-field
django-storages
djangorestframework
djangorestframework-simplejwt
google-api-core
google-auth
google-cloud-core
google-cloud-storage
google-resumable-media
googleapis-common-protos
googlemaps
psycopg2-binary
jsonschema
pylint
autopep8
PyJWT
python-dotenv
pytz
drf_spectacular
pillow
